+++ biom/table.py
@@ -447,6 +447,32 @@ class Table(object):
         """The sparse matrix object"""
         return self._data
 
+    def length(self, axis='sample'):
+        """Return the length of an axis
+
+        Parameters
+        ----------
+        axis : {'sample', 'observation'}, optional
+            The axis to operate on
+
+        Raises
+        ------
+        UnknownAxisError
+            If provided an unrecognized axis.
+
+        Examples
+        --------
+        >>> from biom import example_table
+        >>> print example_table.length(axis='sample')
+        3
+        >>> print example_table.length(axis='observation')
+        2
+        """
+        if axis not in ('sample', 'observation'):
+            raise UnknownAxisError(axis)
+
+        return self.shape[1] if axis == 'sample' else self.shape[0]
+
     def add_group_metadata(self, group_md, axis='sample'):
         """Take a dict of group metadata and add it to an axis
 
@@ -2625,11 +2651,19 @@ class Table(object):
         generator
             Yields ``(observation_id, sample_id)`` for each nonzero element
         """
-        # this is naively implemented. If performance is a concern, private
-        # methods can be written to hit against the underlying types directly
-        for o_idx, samp_vals in enumerate(self.iter_data(axis="observation")):
-            for s_idx in samp_vals.nonzero()[0]:
-                yield (self.ids(axis='observation')[o_idx], self.ids()[s_idx])
+        self._data.tocsr()
+        samp_ids = self.ids()
+        obs_ids = self.ids(axis='observation')
+
+        indptr = self._data.indptr
+        indices = self._data.indices
+        for row_idx in range(indptr.size - 1):
+            start = indptr[row_idx]
+            end = indptr[row_idx+1]
+
+            obs_id = obs_ids[row_idx]
+            for col_idx in indices[start:end]:
+                yield (obs_id, samp_ids[col_idx])
 
     def nonzero_counts(self, axis, binary=False):
         """Get nonzero summaries about an axis